fixed VisMetadata location
